refactor: 코드정리 (usedailydata 생성)
