+++ src/components/Calendar.jsx
@@ -3,17 +3,14 @@ import './styles/Calendar.scss';
 import montharrowleft from '../picture/montharrleft.png';
 import montharrowright from '../picture/montharrright.png';
 import Sample from '../picture/sample.svg';
+import UseDailyData, { kcal } from '../components/UseDailyData.jsx'; // 훅 가져오기
 
-const Calendar = ({
-  checkKcal,
-  checkExercise,
-  selectDate,
-  currentYearMonth,
-}) => {
+const Calendar = ({ selectDate }) => {
+  const { checkKcal, checkExercise } = UseDailyData();
   const daysOfWeek = ['일', '월', '화', '수', '목', '금', '토'];
 
-  const [currentYear, setCurrentYear] = useState(currentYearMonth.year);
-  const [currentMonth, setCurrentMonth] = useState(currentYearMonth.month);
+  const [currentYear, setCurrentYear] = useState(new Date().getFullYear());
+  const [currentMonth, setCurrentMonth] = useState(new Date().getMonth() + 1);
 
   const getFirstDayOfMonth = () => {
     return new Date(currentYear, currentMonth - 1, 1).getDay();
@@ -49,11 +46,9 @@ const Calendar = ({
     return `${year}-${('0' + month).slice(-2)}-${('0' + day).slice(-2)}`;
   };
 
-  // 현재 월은 currentYearMonth가 변경될 때 설정되어야 한다.
   useEffect(() => {
-    setCurrentYear(currentYearMonth.year);
-    setCurrentMonth(currentYearMonth.month);
-  }, [currentYearMonth]);
+    // 현재 월이 변경되면 뭔가 작업이 필요하면 여기에 추가
+  }, [currentYear, currentMonth]);
 
   return (
     <div className="calendar">
